Rename pathfinding_util.py -> util.py
